refactor: move emitTypeSpecHeader to emit.ts
